% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbt.complexity.R
\name{gbt.complexity}
\alias{gbt.complexity}
\title{Return complexity of model in terms of hyperparameters.}
\usage{
gbt.complexity(model, type)
}
\arguments{
\item{model}{object or pointer to object of class \code{ENSEMBLE}}

\item{type}{currently supports "xgboost" or "lightgbm"}
}
\value{
\code{list} with \code{type} hyperparameters.
}
\description{
\code{gbt.complexity} creates a \code{data.frame} of feature importance in a model
}
\details{
Returns the complexity of \code{model} in terms of hyperparameters associated
to model \code{type}.
}
\examples{
\donttest{
## Load data
data(caravan.train, package = "agtboost")
train <- caravan.train
mod <- gbt.train(train$y, train$x, loss_function = "logloss", verbose=10)
feature_names <- colnames(train$x)
imp <- gbt.importance(feature_names, mod)
imp
}

}
