% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbt.pred.R
\name{gbt.pred}
\alias{gbt.pred}
\title{GBTorch Prediction}
\usage{
gbt.pred(object, newdata)
}
\arguments{
\item{object}{Object or pointer to object of class \code{ENSEMBLE}}

\item{newdata}{Design matrix of data to be predicted. Type \code{matrix}}
}
\value{
For regression or binary classification, it returns a vector of length \code{nrows(newdata)}.
}
\description{
\code{gbt.pred} is an interface for predicting from a \code{gbtorch} model.
}
\details{
The prediction function for \code{gbtorch}.
Using the generic \code{predict} function in R is also possible, using the same arguments.
}
\examples{
## A simple gtb.train example with linear regression:
x <- runif(500, 0, 4)
y <- rnorm(500, x, 1)
x.test <- runif(500, 0, 4)
y.test <- rnorm(500, x.test, 1)

mod <- gbt.train(y, as.matrix(x))
y.pred.1 <- gbt.pred( mod, as.matrix( x.test ) )

## predict i overloaded
y.pred.2 <- predict( mod, as.matrix( x.test ) )

plot(x.test, y.test)
points(x.test, y.pred.1, col="red")
points(x.test, y.pred.2, col="blue")


}
\references{
B. Å. S. Lunde, T. S. Kleppe and H. J. Skaug, "An information criterion for gradient boosted trees"
publishing details, \url{}
}
\seealso{
\code{\link{gbt.train}}
}
