% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbt.save.R
\name{gbt.save}
\alias{gbt.save}
\title{Save a GBTorch Model}
\usage{
gbt.save(gbt_model, file)
}
\arguments{
\item{gbt_model}{Model object or pointer to object of class \code{ENSEMBLE}}

\item{file}{Valid file-path}
}
\value{
Txt file that can be loaded using \code{gbt.load}.
}
\description{
\code{gbt.save} is an interface for storing a \code{gbtorch} model.
}
\details{
The model-storage function for \code{gbtorch}.
Saves a GBT model a a txt file. Might be retrieved using \code{gbt.load}
}
\seealso{
\code{\link{gbt.load}}
}
