% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbt.pred.R
\name{predict.Rcpp_GBT_COUNT_AUTO}
\alias{predict.Rcpp_GBT_COUNT_AUTO}
\title{GBTorch Count-Regression Auto Prediction}
\usage{
\method{predict}{Rcpp_GBT_COUNT_AUTO}(object, newdata, ...)
}
\arguments{
\item{object}{Object or pointer to object of class \code{GBT_ZI_MIX}}

\item{newdata}{Design matrix of data to be predicted. Type \code{matrix}}

\item{...}{additional parameters passed. Currently not in use.}
}
\value{
For regression or binary classification, it returns a vector of length \code{nrows(newdata)}.
}
\description{
\code{predict} is an interface for predicting from a \code{gbtorch} model.
}
\details{
The prediction function for \code{gbtorch}.
Using the generic \code{predict} function in R is also possible, using the same arguments.
}
\examples{
## A simple gtb.train example with linear regression:
## Random generation of zero-inflated poisson
2+2

}
\references{
B. Å. S. Lunde, T. S. Kleppe and H. J. Skaug, "An information criterion for gradient boosted trees"
publishing details,
}
\seealso{
\code{\link{gbt.train}}
}
