% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbt.pred.R
\name{predict.Rcpp_GBT_ZI_MIX}
\alias{predict.Rcpp_GBT_ZI_MIX}
\title{GBTorch Zero-Inflated Mixture Prediction}
\usage{
\method{predict}{Rcpp_GBT_ZI_MIX}(object, newdata, ...)
}
\arguments{
\item{object}{Object or pointer to object of class \code{GBT_ZI_MIX}}

\item{newdata}{Design matrix of data to be predicted. Type \code{matrix}}

\item{...}{additional parameters passed. Currently not in use.}
}
\value{
For regression or binary classification, it returns a vector of length \code{nrows(newdata)}.
}
\description{
\code{predict} is an interface for predicting from a \code{gbtorch} model.
}
\details{
The prediction function for \code{gbtorch}.
Using the generic \code{predict} function in R is also possible, using the same arguments.
}
\examples{
## A simple gtb.train example with linear regression:
## Random generation of zero-inflated poisson
rzip <- function(n, lambda, prob){
bin <- rbinom(n, 1, prob)
y <- numeric(n)
for(i in 1:n){
    y[i] <- ifelse(bin[i]==1, 0, rpois(1, lambda))
}
return(y)
}

prob <- 0.4
lambda <- 3
y <- rzip(1000, lambda, prob)
x <- as.matrix(runif(1000, 1,5))

mod <- gbt.train(y, x, loss_function = "zero_inflation::poisson", verbose=1)
pred <- predict(mod, x)
range(pred) # mean predictions
(1-prob)*lambda # true mean

plot(y, pred, ylim=c(0,max(y)))

}
\references{
B. Å. S. Lunde, T. S. Kleppe and H. J. Skaug, "An information criterion for gradient boosted trees"
publishing details,
}
\seealso{
\code{\link{gbt.train}}
}
