% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbt.load.R
\name{gbt.load}
\alias{gbt.load}
\title{Load a GBTorch Model}
\usage{
gbt.load(file)
}
\arguments{
\item{file}{Valid file-path to a stored GBTorch model}
}
\value{
Trained GBTorch model.
}
\description{
\code{gbt.load} is an interface for loading a \pkg{agtboost} model.
}
\details{
The load function for \pkg{agtboost}.
Loades a GBT model from a txt file.
}
\seealso{
\code{\link{gbt.save}}
}
